+++ datahub/dataset/interaction/views.py
@@ -6,7 +6,7 @@ from datahub.core.query_utils import (
     get_front_end_url_expression,
 )
 from datahub.dataset.core.views import BaseDatasetView
-from datahub.interaction.models import Interaction, InteractionExportCountry
+from datahub.interaction.models import Interaction
 from datahub.interaction.queryset import get_base_interaction_queryset
 from datahub.metadata.query_utils import get_sector_name_subquery, get_service_name_subquery
 
@@ -68,23 +68,4 @@ class InteractionsDatasetView(BaseDatasetView):
             'service_delivery',
             'subject',
             'theme',
-        )
-
-
-class InteractionsExportCountryDatasetView(BaseDatasetView):
-    """
-    A GET API view to return all interaction data related to export country
-     as required for syncing by Data-flow periodically.
-    """
-
-    def get_dataset(self):
-        """Returns list of company_export_country_history records"""
-        return InteractionExportCountry.objects.values(
-            'country__name',
-            'country__iso_alpha2_code',
-            'created_on',
-            'id',
-            'interaction__company_id',
-            'interaction__id',
-            'status',
         )
